+++ cni/plugin.go
@@ -86,7 +86,7 @@ func (plugin *Plugin) Execute(api PluginApi) (err error) {
 	pluginInfo := cniVers.PluginSupports(supportedVersions...)
 
 	// Parse args and call the appropriate cmd handler.
-	cniErr := cniSkel.PluginMainWithError(api.Add, api.Delete, pluginInfo)
+	cniErr := cniSkel.PluginMainWithError(api.Add, api.Get, api.Delete, pluginInfo, plugin.version)
 	if cniErr != nil {
 		cniErr.Print()
 		return cniErr
@@ -105,7 +105,7 @@ func (plugin *Plugin) DelegateAdd(pluginName string, nwCfg *NetworkConfig) (*cni
 
 	os.Setenv(Cmd, CmdAdd)
 
-	res, err := cniInvoke.DelegateAdd(pluginName, nwCfg.Serialize())
+	res, err := cniInvoke.DelegateAdd(pluginName, nwCfg.Serialize(), nil)
 	if err != nil {
 		return nil, fmt.Errorf("Failed to delegate: %v", err)
 	}
@@ -127,7 +127,7 @@ func (plugin *Plugin) DelegateDel(pluginName string, nwCfg *NetworkConfig) error
 
 	os.Setenv(Cmd, CmdDel)
 
-	err = cniInvoke.DelegateDel(pluginName, nwCfg.Serialize())
+	err = cniInvoke.DelegateDel(pluginName, nwCfg.Serialize(), nil)
 	if err != nil {
 		return fmt.Errorf("Failed to delegate: %v", err)
 	}
